Refactor indirect_calls_executor
