+++ core/parentchain/indirect-calls-executor/src/executor/call_worker.rs
@@ -1,0 +1,73 @@
+// Copyright 2020-2023 Litentry Technologies GmbH.
+// This file is part of Litentry.
+//
+// Litentry is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Litentry is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Litentry.  If not, see <https://www.gnu.org/licenses/>.
+
+use crate::{error::Error, executor::Executor, IndirectCallsExecutor};
+use itp_node_api::{
+	api_client::ParentchainUncheckedExtrinsic,
+	metadata::{
+		pallet_teerex::TeerexCallIndexes, provider::AccessNodeMetadata, Error as MetadataError,
+		NodeMetadataTrait,
+	},
+};
+use itp_sgx_crypto::{key_repository::AccessKey, ShieldingCryptoDecrypt, ShieldingCryptoEncrypt};
+use itp_stf_executor::traits::StfEnclaveSigning;
+use itp_top_pool_author::traits::AuthorApi;
+use itp_types::{CallWorkerFn, H256};
+
+pub struct CallWorker {}
+
+impl<ShieldingKeyRepository, StfEnclaveSigner, TopPoolAuthor, NodeMetadataProvider>
+	Executor<ShieldingKeyRepository, StfEnclaveSigner, TopPoolAuthor, NodeMetadataProvider>
+	for CallWorker
+where
+	ShieldingKeyRepository: AccessKey,
+	<ShieldingKeyRepository as AccessKey>::KeyType: ShieldingCryptoDecrypt<Error = itp_sgx_crypto::Error>
+		+ ShieldingCryptoEncrypt<Error = itp_sgx_crypto::Error>,
+	StfEnclaveSigner: StfEnclaveSigning,
+	TopPoolAuthor: AuthorApi<H256, H256> + Send + Sync + 'static,
+	NodeMetadataProvider: AccessNodeMetadata,
+	NodeMetadataProvider::MetadataType: NodeMetadataTrait,
+{
+	type Call = CallWorkerFn;
+
+	fn call_index(&self, call: Self::Call) -> [u8; 2] {
+		call.0
+	}
+
+	fn call_index_from_metadata(
+		&self,
+		metadata_type: &NodeMetadataProvider::MetadataType,
+	) -> Result<[u8; 2], MetadataError> {
+		metadata_type.call_worker_call_indexes()
+	}
+
+	fn execute(
+		&self,
+		context: &IndirectCallsExecutor<
+			ShieldingKeyRepository,
+			StfEnclaveSigner,
+			TopPoolAuthor,
+			NodeMetadataProvider,
+		>,
+		extrinsic: ParentchainUncheckedExtrinsic<Self::Call>,
+	) -> Result<(), Error> {
+		let (_, request) = extrinsic.function;
+		let (shard, cypher_text) = (request.shard, request.cyphertext);
+		log::debug!("Found trusted call extrinsic, submitting it to the top pool");
+		context.submit_trusted_call(shard, cypher_text);
+		Ok(())
+	}
+}
